requirements.txt
oandapyv20==0.7.2
pandas==2.2.3
python-dotenv[cli]==1.1.0
ta==0.11.0
